code redundancy update and edit options
removed all redundant code for the current forms components and re-organised the file structure to store things appropriately, disabled editing for formbuilder
